agent/pyproject.toml
[tool.poetry]
name = "scifly"
version = "0.1.0"
description = ""
authors = ["Bilal"]
readme = "README.md"
packages = [{include = "sample_agent"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
copilotkit = "0.1.55"
langchain = "0.3.21"
langgraph = "0.3.27"
langsmith = "0.3.45"
openai = "^1.68.2"
fastapi = "^0.115.5"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.0"
langchain-groq = ">=0.2.3"
langchain-google-genai = ">=2.0.5"
langchain-openai = ">=0.2.3"
langgraph-cli = {version = "0.3.4", extras = ["inmem"]}
ipython = "^8.26.0"
google-genai ="1.28.0"

[tool.poetry.scripts]
demo = "sample_agent.demo:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"